Using uint64_t for x generation #39

+++ include/nil/crypto3/zk/commitments/detail/polynomial/basic_fri.hpp
@@ -221,7 +221,7 @@ namespace nil {
 
                             // TODO: how to sample x?
                             std::size_t domain_size = fri_params.D[0]->m;
-                            std::size_t x_index = (transcript.template int_challenge<std::size_t>())%domain_size;
+                            std::size_t x_index = (transcript.template int_challenge<std::uint64_t>())%domain_size;
                             std::size_t x_next_index;
                             typename FieldType::value_type x =
                                 fri_params.D[0]->get_domain_element(1).pow(x_index);
